feat(cam): move generation controls into the CAM panel gear (#675)
Status, Stop/Resume/Retry and the backend choice now sit behind a gear in the
CAM panel header instead of the status bar. That is where generation is
visible — the operations being generated are listed directly below it.

All of it moved rather than just the backend picker. A backend menu in the CAM
panel with a Stop button still down in the status bar would be two homes for
one concern, which is worse than either place alone.

A menu hides what is inside it, so two things compensate: the gear marks itself
when generation is paused or has failed, and it carries the status in its
accessible name — so a screen reader, and the e2e suite, can read it without
opening anything. The per-operation pause badges in the tree say the same at
row level.

The status-bar control is gone and `StatusBarExtras` is back to just the depth
legend.
