+++ src/components/INDEX.md
@@ -24,8 +24,8 @@ React UI. Components are organized by feature area. Plain CSS for styling — no
 - `cam/` — CAM panels (tools, operations, parameters); per-parameter reference icons via `OperationParameterReference`. `ToolLibraryDialog` renders the focused library-import modal; the pure filtering/selection helpers live in `toolLibraryDialogModel.ts` with their own structural tests. `operationTargetLists.ts` computes the context menu's "Add/Remove to operation" candidates against the store's own `isOperationTargetValid` guard. `SurfaceSlopeFields.tsx` provides the optional min/max CL-surface slope controls for 3D finish operations. `operationFields.ts` declares the operation panel's field order and grouping as data — one ordered table of `{ id, group, paramRef, appliesTo }` that `CAMPanel` walks — so a control's position is a property of the field rather than of the JSX branch it was written into (#559). Which patterns the `pattern` dropdown offers is no longer the renderer's own list: both it and `operationFields.ts`'s applicability read `engine/toolpaths/pocketPatterns.ts`, the same table the generators dispatch on, so a kind can never be offered a pattern its generator does not implement (#609). `OperationAddMenu` partitions the Add menu on the same validity hint the rows already carried: kinds the selection can take render first, the rest collapse into one counted `DisclosureSection`, and an empty selection states its single precondition once instead of eleven times (#732).
 - `feature-tree/` — sketch feature tree UI (reordering, visibility, grouping) plus `PropertiesPanel`, `ZRangeSlider` (generalized single/bulk Z-range with locked handles), `BulkTabProperties`/`BulkClampProperties` (bulk-edit panels with center-preserving dims and mixed-value display), `DraftTextInput`/`DraftNumberInput` (shared inline-edit fields), `mixedValue.ts` (pure common/mixed aggregation helpers), the extracted feature/tab/clamp context menu, and `ModelOrientationSection` (post-import 3D orientation + lift controls for imported models)
 - `layout/` — app shell (toolbars, sidebars, mode switching), including the shared `AppearanceControl` and `LanguageControl` (interface-language selector backed by `src/i18n/`); toolbar internals are documented in `layout/toolbar/INDEX.md`
-- `layout/GenerationStatusControl.tsx` — generation status, Stop/Resume/Retry, and the execution-backend choice (issue #675). Small by intent: the CAM panel is not redesigned, this answers "is it working, and can I stop it" — questions the app could not previously answer, because generation blocked the thread that would have drawn the answer. Stop is **disabled** under the main-thread backend rather than offered and quietly ineffective, and the menu says so
-- `layout/StatusBarExtras.tsx` — what the status bar shows beyond its own toggles: the collapsed feature-colour legend and generation status. Extracted from `App.tsx` when generation status made that slot a composite
+- `cam/GenerationSettingsMenu.tsx` — generation status, Stop/Resume/Retry and the execution-backend choice, behind the gear in the CAM panel header (issue #675). Started in the status bar and moved here: this is where generation is visible, with the generating operations listed directly below. All of it sits behind one gear rather than split across two surfaces — a backend picker in the panel and a Stop button in the status bar would be two homes for one concern. Because a menu hides its contents, the gear marks itself when generation is paused or failed, and carries the status in its accessible name so a screen reader gets it without opening anything. Stop is **disabled** under the main-thread backend rather than offered and quietly ineffective, and says why
+- `layout/StatusBarExtras.tsx` — what the status bar shows beyond its own toggles; now just the collapsed feature-colour legend
 - `project/` — project-level UI (new/open/save, stock, machine, units), including `UnitConversionDialog` and `UnitConversionContext` for the shared explicit convert-vs-reinterpret units decision. Import dialog is `ImportGeometryDialog` with analysis delegated to `useImportGeometryAnalysis` (parse/classify caching hook), `ImportGeometryModeSection` (mode select + summary), and `importModelFile` (3D model import utility). `importedModelArtifacts.ts` owns the silhouette/profile/top-view derivation shared by first import and post-import re-orientation, plus the rigid `reorientImportedModel` and non-deforming `liftImportedModel` commits (issue #241).
 - `export/` — export dialogs: G-code (`ExportDialog`, with a per-operation checklist backed by the pure helpers in `exportOperationSelection.ts`, plus an "Inspect exported motion" action for a single eligible operation that opens `ExportedMotionDebugDialog` — issue #356), the Export Model dialog (`ModelExportDialog`, STL mesh + 2D design SVG via `src/engine/modelExport/`), and the Print Design dialog (`PrintDesignDialog`, backed by `src/engine/designPrint/`)
 - `machine/` — machine definition editor (focused form + raw JSON + Zod validation)
