+++ src/styles/layout.css
@@ -494,13 +494,12 @@
 }
 
 /*
- * Portalled to the body and positioned inline (see GenerationStatusControl):
- * the status bar is at the bottom of the window and below the centre stage in
- * stacking order, so an in-flow menu either fell off the fold or rendered
- * underneath the viewport overlay. Only the visual styling is inherited from
- * `.appearance-menu`; placement comes from `usePortalPosition`.
+ * Portalled to the body and positioned inline (see GenerationSettingsMenu): the
+ * CAM panel scrolls and carries its own stacking context, so an in-flow menu
+ * would be clipped by the panel it hangs off. Only the visual styling comes
+ * from `.appearance-menu`; placement comes from `usePortalPosition`.
  */
-.generation-status__menu {
+.cam-generation-menu {
   right: auto;
   width: 300px;
 }
@@ -510,18 +509,38 @@
  * (swatch / copy / check) would give the label a 38px column to wrap in. Two
  * columns instead, so the copy gets the width.
  */
-.generation-status__menu .appearance-menu__option {
+.cam-generation-menu .appearance-menu__option {
   grid-template-columns: 1fr 20px;
 }
 
-.generation-status {
-  gap: 6px;
-}
-
-.generation-status__summary {
+.cam-generation-menu__status {
+  padding: 2px 8px 8px;
   color: var(--text-dim);
   font-size: 11px;
-  white-space: nowrap;
+}
+
+.cam-generation-menu__actions {
+  display: flex;
+  gap: 6px;
+  padding: 0 4px 8px;
+  border-bottom: 1px solid var(--line);
+  margin-bottom: 8px;
+}
+
+.cam-generation-menu__actions .cam-header-action {
+  display: inline-flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.cam-generation-menu__actions svg {
+  width: 12px;
+  height: 12px;
+}
+
+/* A menu hides its contents, so a stopped or failed pipeline marks the gear. */
+.cam-generation-gear--flagged {
+  color: var(--accent-strong);
 }
 
 .appearance-menu__heading {
